Refactor currency moving out of google module all convert currency methods.
